fix: Added init method in Redis Util

+++ src/utils/redis.util.ts
@@ -2,15 +2,24 @@ import { Redis } from "ioredis";
 import { CustomError } from "./error.util";
 
 export class RedisUtil {
-  private client: Redis;
+  private client: Redis | undefined;
   private errCode = 501;
   private errName = "Redis Error";
 
   constructor(client: Redis) {
+    this.init(client);
+  }
+
+  private async init(client: Redis) {
+    await client.connect();
     this.client = client;
   }
 
   async get<T>(key: string): Promise<T | null> {
+    if (!this.client) {
+      throw new CustomError(this.errName, "Client error", this.errCode);
+    }
+
     const data = await this.client.get(key);
 
     if (data) {
@@ -21,11 +30,18 @@ export class RedisUtil {
   }
 
   async set<T>(key: string, data: T): Promise<void> {
+    if (!this.client) {
+      throw new CustomError(this.errName, "Client error", this.errCode);
+    }
+
     const stringifiedData = JSON.stringify(data);
     await this.client.set(key, stringifiedData);
   }
-
+  
   async update<I, T>(key: string, data: T): Promise<void> {
+    if (!this.client) {
+      throw new CustomError(this.errName, "Client error", this.errCode);
+    }
     const record = await this.get<any>(key);
 
     if (!record) {
@@ -42,6 +58,9 @@ export class RedisUtil {
   }
 
   async delete(key: string): Promise<void> {
+    if (!this.client) {
+      throw new CustomError(this.errName, "Client error", this.errCode);
+    }
     await this.client.del(key);
   }
 }
